:bug: Fix free size column

+++ database/migrations/2023_08_19_091002_create_storages_table.php
@@ -26,7 +26,7 @@ return new class extends Migration {
                 $table->index('created_at');
                 $table->index('updated_at');
 
-                $table->integer('free_size')
+                $table->bigInteger('free_size')
                     ->storedAs('total_size - used_size');
             }
         );
